+++ changelog.md
@@ -1,6 +1,19 @@
 ## 0.0.6
 
 * Changed base income from $0.50 - $3.75 to $2.50 - $10.75
+## 0.0.6
+
+Added: 
+* Added buy command for Gas Station and Retail business types
+* Added upgrades for the Retail business type
+* Added Profile Command
+
+Changes:
+* Changed the start command to notify the user if they already have a profile set up
+
+Fixed:
+* Corrected Retail business type to Gas Station
+* Fixed buy command not truncating final cost down to 2 decimals
 
 ## 0.0.5